+++ src/components/data-management/FileUploadModal.jsx
@@ -20,7 +20,7 @@ import { CheckCircleTwoTone, CloseCircleTwoTone, DeleteOutlined } from '@ant-des
 import Dropzone from 'react-dropzone';
 import techOptions from '../../utils/fileUploadSpecifications';
 import UploadStatus from '../../utils/UploadStatus';
-
+import checkIfFileValid from '../../utils/checkIfFileValid';
 import pushNotificationMessage from '../../utils/pushNotificationMessage';
 
 const { Text, Title, Paragraph } = Typography;
@@ -43,7 +43,7 @@ const FileUploadModal = (props) => {
   const onDrop = (acceptedFiles) => {
     const newList = [];
 
-    const acceptedFilesRegexp = `(${techOptions[selectedTech].acceptedFiles.join('|')})$`;
+    // const acceptedFilesRegexp = `(${techOptions[selectedTech].acceptedFiles.join('|')})$`;
 
     let filesNotInFolder = false;
     const filteredFiles = acceptedFiles
@@ -65,28 +65,28 @@ const FileUploadModal = (props) => {
     filteredFiles.forEach((file) => {
       let fileName = null;
       const error = [];
-
       // First character of file.path === '/' means a directory is uploaded
       // Remove initial slash so that it does not create an empty directory in S3
       const paths = file.path.split('/');
       fileName = `${paths[paths.length - 2]}/${paths[paths.length - 1]}`;
+      const valid = checkIfFileValid(fileName, selectedTech);
 
-      const isValidType = (
-        techOptions[selectedTech].validMimeTypes
-          .includes(
-            mime.lookup(file.path),
-          )
-        || techOptions[selectedTech].validExtensionTypes
-          .includes(
-            path.extname(file.path),
-          )
-      );
+      // const isValidType = (
+      //   techOptions[selectedTech].validMimeTypes
+      //     .includes(
+      //       mime.lookup(file.path),
+      //     )
+      //   || techOptions[selectedTech].validExtensionTypes
+      //     .includes(
+      //       path.extname(file.path),
+      //     )
+      // );
 
-      if (!isValidType) error.push('Invalid file type.');
+      if (!valid.isValidType) error.push('Invalid file type.');
 
-      const acceptedFilenames = new RegExp(acceptedFilesRegexp, 'gi');
-      const isValidFilename = fileName.match(acceptedFilenames) !== null;
-      if (!isValidFilename) error.push('Invalid file name.');
+      // const acceptedFilenames = new RegExp(acceptedFilesRegexp, 'gi');
+      // const isValidFilename = fileName.match(acceptedFilenames) !== null;
+      if (!valid.isValidFilename) error.push('Invalid file name.');
 
       newList.push({
         name: fileName,
@@ -95,7 +95,7 @@ const FileUploadModal = (props) => {
           status: UploadStatus.UPLOADING,
           progress: 0,
         },
-        valid: isValidType && isValidFilename,
+        valid: valid.isValidType && valid.isValidFilename,
         errors: error.join(', '),
       });
     });
