+++ src/redux/actions/samples/updateSampleFile.js
@@ -8,6 +8,7 @@ import endUserMessages from '../../../utils/endUserMessages';
 import pushNotificationMessage from '../../../utils/pushNotificationMessage';
 import mergeObjectWithArrays from '../../../utils/mergeObjectWithArrays';
 import UploadStatus from '../../../utils/UploadStatus';
+import checkIfFileValid from '../../../utils/checkIfFileValid';
 
 const updateSampleFile = (
   sampleUuid,
@@ -16,11 +17,15 @@ const updateSampleFile = (
 ) => async (dispatch, getState) => {
   const updatedAt = moment().toISOString();
   const sample = getState().samples[sampleUuid];
+  // we'll need to remove the hard-coded 10x tech type once we start
+  // supporting other types and save the chosen tech type in redux
+
+  const valid = checkIfFileValid(fileName, '10X Chromium');
+
   try {
     // Save sample only if upload is successful or error
     if (fileDiff.upload.status === UploadStatus.UPLOADED
       || fileDiff.upload.status === UploadStatus.UPLOAD_ERROR) {
-      console.log('FILENAMES IS ', sample.fileNames);
       const diffObject = {
         fileNames: sample.fileNames,
         files: {
@@ -28,6 +33,7 @@ const updateSampleFile = (
           [fileName]: {
             ...sample.files[fileName],
             ...fileDiff,
+            valid: valid.isValidFilename && valid.isValidType,
           },
         },
       };
